+++ bundle/edu.gemini.seqexec.odb/src/main/scala/edu/gemini/seqexec/odb/SeqFailure.scala
@@ -1,4 +1,4 @@
-package edu.gemini.seqexec.shared
+package edu.gemini.seqexec.odb
 
 import edu.gemini.pot.sp.SPObservationID
